# PERF: Improve intrinsics for tobits and pack on Apple silicon