@charris 2022-08-09
This broke our aarch64 wheel builds, see https://app.travis-ci.com/github/MacPython/numpy-wheels/jobs/579262744. We may need to revert this.